finish admin functions

+++ src/main/java/softuniBlog/controller/ArticleController.java
@@ -2,6 +2,7 @@ package softuniBlog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -59,6 +60,13 @@ public class ArticleController {
             return "redirect:/";
         }
 
+        if (!(SecurityContextHolder.getContext().getAuthentication()instanceof AnonymousAuthenticationToken)) {
+            UserDetails principal = (UserDetails) SecurityContextHolder.getContext()
+                    .getAuthentication().getPrincipal();
+            User entityUser = this.userRepository.findByEmail(principal.getUsername());
+
+            model.addAttribute("user", entityUser);
+        }
         Article article = this.articleRepository.findOne(id);
 
         model.addAttribute("article", article);
@@ -76,6 +84,10 @@ public class ArticleController {
         }
         Article article = this.articleRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article" + id;
+        }
+
         model.addAttribute("view", "article/edit");
         model.addAttribute("article", article);
 
@@ -91,6 +103,10 @@ public class ArticleController {
         }
 
         Article article = this.articleRepository.findOne(id);
+
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
 
         article.setTitle(articleBindingModel.getTitle());
         article.setContent(articleBindingModel.getContent());
@@ -110,6 +126,10 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article" + id;
+        }
+
         model.addAttribute("article", article);
         model.addAttribute("view", "article/delete");
 
@@ -126,8 +146,22 @@ public class ArticleController {
 
         Article article = this.articleRepository.findOne(id);
 
+        if (!isUserAuthorOrAdmin(article)) {
+            return "redirect:/article/" + id;
+        }
+
         this.articleRepository.delete(article);
 
         return "redirect:/";
+    }
+
+    private boolean isUserAuthorOrAdmin(Article article) {
+
+        UserDetails user = (UserDetails) SecurityContextHolder.getContext()
+                .getAuthentication().getPrincipal();
+
+        User userEntity = this.userRepository.findByEmail(user.getUsername());
+
+        return userEntity.isAdmin() || userEntity.isAuthor(article);
     }
 }
